moving state in About component back up to router
